Move unittests to dalitests
